backend/requirements.txt
django
djangorestframework
markdown
django-filter
dj-rest-auth
drf-spectacular